Update README.md
The Cakeshop project is an online platform for customers to browse, customize, and order cakes. It includes user registration, a product catalog, real-time customization, and secure order management. Admins can manage products and orders through an admin panel. The frontend, built with HTML, CSS, and JavaScript, is mostly complete, and work is underway on integrating React and developing the backend using Java or Python with MySQL. The project prioritizes user-friendly design, data security, and efficient order processing.